Avoid conflicts when same rating HTML is repeated on page
When the same rating HTML is repeated more than once on the page (it may
happen as the product is listed both on the category page and among the
recently viewed items) then setting `eid` from `name` attribute will not
work correctly, as it will catch 2 ratings instead of only one.

Adding a unique counter to the element parent and adding it to `eid`
prevents this kind of issue.